requirements.txt
streamlit
numpy  
tensorflow
Pillow
streamlit_drawable_canvas
